Moved db connection methods to utils folder
