+++ src/utils/index.ts
@@ -2,5 +2,6 @@ export * from './api.utils';
 export * from './array.utils';
 export * from './cache.utils';
 export * from './crypto.utils';
+export * from './db.utils';
 export * from './middleware.utils';
 export * from './server.utils';
